+++ CHANGELOG.md
@@ -4,12 +4,19 @@ Changes in a project.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.0.4]
 
 ### Added
 
 - Basic docs with library interface reference
-- Deleting tasks' attachment
+
+### Changed
+
+- Switching to the new ToDo API
+
+### Removed
+
+- `Attachment`, `Sensitivity` as not supported by the new API
 
 ## [0.0.3] - 2020-06-20